\name{get.changescore}
\alias{get.changescore}
\title{Compute the changescore for the given networkcs object and
the provided edge toggles.}
\usage{
  get.changescore(ncs, toggle.edges)
}
\arguments{
  \item{ncs}{networkcs object}

  \item{toggle.edges}{M x 2 matrix of edge toggles to
  perform}
}
\value{
  a vector of statistics (with an element for each term in
  the provided model.
}
\description{
  (description will go here)
}
\author{
  Chris DuBois
}

