\name{network.for.changescores}
\alias{network.for.changescores}
\title{Create an object that allows for online updating of network statistics.}
\usage{
  network.for.changescores(object)
}
\arguments{
  \item{object}{ergm.formula object}
}
\value{
  a networkcs object (short for "changescore")
}
\description{
  (decription will go here)
}
\author{
  Chris DuBois
}

